Initial pass in implementation of writing files.
server.cc, and BufferedReader currently implementing just saving files from a
socket to the disk. These have been copied to the tests/ subfolder so that they
can be used as a jumping off point for testing scalability in a general sense.

Updates have also been made to fileuv, ccuv, and netuv that were necessary to
make these changes.